% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SevereInHospitalizedPatientsValidation.R
\docType{package}
\name{SevereInHospitalizedPatientsValidation}
\alias{SevereInHospitalizedPatientsValidation}
\title{SevereInHospitalizedPatientsValidation}
\description{
SevereInHospitalizedPatientsValidation
}
